pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nqmp-bench"
version = "0.1.0"
description = "Negation & Quantifier Minimal Pairs — tiny benchmark harness"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Sash Sinha"}]
dependencies = [
  "requests>=2.31",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "matplotlib>=3.8",
  "tabulate>=0.9",
  "ruff>=0.5.0",
]

[project.urls]
homepage = "https://github.com/sashsinha/nqmp-bench"

[project.scripts]
nqmp = "nqmp_bench.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 80
target-version = "py310"
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "Q"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
docstring-quotes = "double"
multiline-quotes = "double"
